// src/Models/collections.ts
export enum Collections {
    Case = 'Case',
    CompanyAuthUserJoin = 'CompanyAuthUserJoin',
    Company = 'Company',
    CompanyUser = 'CompanyUser',
    CaseCheckpoint = 'CaseCheckpoint',
    WorkflowCheckpoint = 'WorkflowCheckpoint',
    CompanyWorkflow = 'CompanyWorkflow',
    PrescriptionTemplate = 'PrescriptionTemplate',
}